Restructure app_api folder for team and match views
