+++ app_api/modules/matches/views.py
@@ -1,0 +1,15 @@
+from rest_framework import viewsets
+
+from app_api.common.serializers import MatchSerializer, MatchDetailSerializer
+from app_prime_league.models import Match
+
+
+class MatchViewSet(viewsets.ReadOnlyModelViewSet):
+    queryset = Match.objects.all()
+    serializer_class = MatchSerializer
+    detail_serializer_class = MatchDetailSerializer
+
+    def get_serializer_class(self):
+        if self.action == 'retrieve':
+            return self.detail_serializer_class
+        return self.serializer_class
