+++ app_api/modules/teams/views.py
@@ -1,19 +1,6 @@
 from rest_framework import viewsets
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from rest_framework.reverse import reverse
-
-from app_api.modules.teams.serializers import TeamSerializer, TeamDetailSerializer, MatchSerializer, \
-    MatchDetailSerializer
-from app_prime_league.models import Team, Match
-
-
-@api_view(['GET'])
-def api_root(request, format=None):
-    return Response({
-        'teams': reverse('team-list', request=request, format=format),
-        'matches': reverse('match-list', request=request, format=format)
-    })
+from app_api.common.serializers import TeamSerializer, TeamDetailSerializer
+from app_prime_league.models import Team
 
 
 class TeamViewSet(viewsets.ReadOnlyModelViewSet):
@@ -26,13 +13,3 @@ class TeamViewSet(viewsets.ReadOnlyModelViewSet):
             return self.detail_serializer_class
         return self.serializer_class
 
-
-class MatchViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = Match.objects.all()
-    serializer_class = MatchSerializer
-    detail_serializer_class = MatchDetailSerializer
-
-    def get_serializer_class(self):
-        if self.action == 'retrieve':
-            return self.detail_serializer_class
-        return self.serializer_class
