+++ app_api/modules/urls.py
@@ -1,7 +1,9 @@
 from django.urls import path
 from rest_framework.routers import SimpleRouter
 
-from app_api.modules.teams.views import TeamViewSet, api_root, MatchViewSet
+from app_api.modules.views import api_root
+from app_api.modules.teams.views import TeamViewSet
+from app_api.modules.matches.views import MatchViewSet
 
 router = SimpleRouter()
 router.register(r'teams', TeamViewSet, basename='team')
